fix after testing via postman

+++ src/main/java/com/ecommerce/staples_clone/repository/ProductCategoryRepository.java
@@ -6,4 +6,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductCategoryRepository extends JpaRepository<ProductCategory, Long> {}
+public interface ProductCategoryRepository extends JpaRepository<ProductCategory, Integer> {}
